+++ ludwig/modules/optimization_modules.py
@@ -20,7 +20,7 @@ import torch
 from marshmallow import fields, ValidationError
 from marshmallow_dataclass import dataclass
 
-from ludwig.marshmallow_schema_utils.utils import (
+from ludwig.marshmallow_utils.schema import (
     BaseMarshmallowConfig,
     create_cond,
     FloatRange,
