+++ tests/integration_tests/test_custom_components.py
@@ -12,7 +12,7 @@ from ludwig.decoders.base import Decoder
 from ludwig.decoders.registry import register_decoder
 from ludwig.encoders.base import Encoder
 from ludwig.encoders.registry import register_encoder
-from ludwig.marshmallow_schema_utils.test_classes import CustomTestCombinerConfig
+from ludwig.marshmallow_utils.test_classes import CustomTestCombinerConfig
 from ludwig.modules.loss_modules import LogitsInputsMixin, register_loss
 from ludwig.modules.metric_modules import LossMetric, register_metric
 from tests.integration_tests.utils import (
