+++ tests/ludwig/combiners/test_combiners.py
@@ -23,7 +23,7 @@ from ludwig.combiners.combiners import (
     TransformerCombiner,
     TransformerCombinerConfig,
 )
-from ludwig.marshmallow_schema_utils.utils import load_config
+from ludwig.marshmallow_utils.schema import load_config
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
